4.21*: Blog List Expansion, step 9

+++ part4/bloglist/controllers/blogs.js
@@ -16,7 +16,6 @@ blogsRouter.post('/', async (request, response) => {
   }
   const user = await User.findById(decodedToken.id)
 
-  //if (usersInDb.length > 0) {
   const blogEntry = new Blog({
     author: body.author,
     url: body.url,
@@ -28,14 +27,24 @@ blogsRouter.post('/', async (request, response) => {
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
   response.status(201).json(savedBlog)
-  //} else {
-  //  response.status(500).send({ error: 'could not find suitable user' })
-  //}
+
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if ( blogToDelete.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'wrong token provided' })
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
